+++ src/main/java/net/jrodolfo/java_evolution/java06/console_api/package-info.java
@@ -1,4 +1,4 @@
 /**
- * Explanatory module for the Java 6 Console API.
+ * Executable examples for the Java 6 Console API.
  */
 package net.jrodolfo.java_evolution.java06.console_api;